+++ src/dlls/objects/524_WM_spiritpl/WM_spiritpl.c
@@ -529,7 +529,7 @@ void WM_spiritplace_init(GameObject* obj, WmSpiritPlaceMapData* placement)
     obj->animEventCallback = WM_spiritplace_SeqFn;
     obj->anim.rotX = (s16)(placement->rotXByte << 8);
     obj->anim.rotY = (s16)(placement->rotYAngle << 8);
-    state->heightOffset = ((f32)placement->heightOffset / 32767.0f) / 100.0f;
+    state->heightOffset = (placement->heightOffset / 32767.0f) / 100.0f;
     state->unk_04 = 0;
     state->unk_08 = 0;
     state->unk_0A = 0;
